requirements.txt
alembic==0.9.8
bugsnag==3.4.2
blinker==1.4
celery==3.1.17
coverage==4.5.1
Flask==0.12.2
Flask-Cors==3.0.4
Flask-JSON==0.3.2
Flask-Script==2.0.6
Flask-GraphQL==1.4.1
Flask-Mail==0.9.1
google-api-python-client==1.6.7
graphene-sqlalchemy==2.0.0
graphene==2.1
imgkit==1.0.1
MarkupSafe==1.0
more-itertools==4.1.0
numpy==1.15.2
opencv-python==3.4.3.18
pandas==0.23.4
Pillow==5.3.0
pdfkit==0.6.1
psycopg2-binary==2.7.4
py==1.5.3
pytest==3.5.0
python-dateutil==2.7.0
PyJWT==1.6.4
redis==2.10.3
nose==1.3.7
python-editor==1.0.3
SQLAlchemy==1.2.6
tox==3.0.0
SQLAlchemy-Utils==0.33.2
virtualenv==15.2.0
Flask-Testing==0.7.1
typing==3.6.4
flake8==3.5.0
coveralls
validators==0.12.4